refactor ui components to export only components
